Fixed a bug causing to mark external urls before actually downloading them,

+++ src/Root/Stopwatch.java
@@ -20,7 +20,7 @@ public class Stopwatch {
      *
      * @return elapsed CPU time (in milliseconds) since the stopwatch was created
      */
-    public double elapsedTime() {
+    public long elapsedTime() {
         long now = System.currentTimeMillis();
         return (now - start);
     }
